* 去掉WebPageImpl::paintToBit里pitch为0返回的安全检查。没啥必要，保持和ewe一致即可 * WebPageImpl::loadURL增加request.setHTTPReferrer，之前遗漏了referrer。另外cef接口里也要添加。下次再和wke一起加上

+++ content/browser/NavigationController.cpp
@@ -66,7 +66,7 @@ void NavigationController::insertOrReplaceEntry(const blink::WebHistoryItem& ite
     switch (type) {
     case blink::WebStandardCommit:
         ++m_currentOffset;
-        ASSERT(0 <= m_currentOffset && m_currentOffset <= m_items.size());
+        ASSERT(0 <= m_currentOffset && m_currentOffset <= (int)m_items.size());
         if (m_currentOffset == m_items.size()) {
             m_items.append(item);
         } else {
